Bugfixing for test23. Test23 is now working.

+++ CS502_Project/fileSystem.h
@@ -17,9 +17,9 @@ void flushDiskContents();
 int createDir(char* directoryName);
 int createFile(char* directoryName);
 int openFile(char* fileName);
-int writeFile(int inode, int logicalBlock, char* writeBuffer);
-int closeFile(int inode);
-int readFile(int inode, int logicalBlock, char* readBuffer);
+int writeFile(long inode, int logicalBlock, char* writeBuffer);
+int closeFile(long inode);
+int readFile(long inode, int logicalBlock, char* readBuffer);
 
 int formattedDisk;
